+++ src/components/traits/i-active-items/i-active-items.ts
@@ -182,6 +182,10 @@ export default abstract class iActiveItems extends iItems {
 
 	/** @see [[iActiveItems.setActive]] */
 	static setActive(ctx: TraitComponent, value: iActiveItems['ActiveProp'], unsetPrevious?: boolean): boolean {
+		if (!this.isActivatable(ctx, value)) {
+			return false;
+		}
+
 		let
 			activeStore = ctx.field.get('activeStore');
 
@@ -279,6 +283,10 @@ export default abstract class iActiveItems extends iItems {
 
 	/** @see [[iActiveItems.toggleActive]] */
 	static toggleActive(ctx: TraitComponent, value: iActiveItems['ActiveProp'], unsetPrevious?: boolean): iActiveItems['Active'] {
+		if (!this.isActivatable(ctx, value)) {
+			return false;
+		}
+
 		const
 			activeStore = ctx.field.get('activeStore');
 
@@ -315,6 +323,22 @@ export default abstract class iActiveItems extends iItems {
 		}
 
 		return ctx.active;
+	}
+
+	/** @see [[iActiveItems.prototype.getItemByValue]] */
+	static getItemByValue(ctx: TraitComponent, value: Item['value']): CanUndef<Item> {
+		return ctx.items?.find((item) => item.value === value);
+	}
+
+	/**
+	 * Checks if item can possibly be active by its value
+	 *
+	 * @param ctx
+	 * @param value
+	 */
+	protected static isActivatable(ctx: TraitComponent, value: Item['value']): boolean {
+		const item = ctx.getItemByValue(value);
+		return item?.activatable !== false;
 	}
 
 	/**
@@ -359,6 +383,15 @@ export default abstract class iActiveItems extends iItems {
 	 * @emits `change(active: unknown)`
 	 */
 	toggleActive(value: this['ActiveProp'], unsetPrevious?: boolean): iActiveItems['Active'] {
+		return Object.throw();
+	}
+
+	/**
+	 * Returns an item by the specified value
+	 *
+	 * @param value
+	 */
+	getItemByValue(value: Item['value']): CanUndef<Item> {
 		return Object.throw();
 	}
 }