+++ src/components/traits/i-active-items/CHANGELOG.md
@@ -9,6 +9,13 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v??.??.?? (2023-??-??)
+
+#### :rocket: New Feature
+
+* Added ability to exclude an item from being activated by specifying `activatable` flag
+* Added `getItemByValue` method
+
 ## v3.41.0 (2023-03-14)
 
 #### :rocket: New Feature